requirements.txt
# Core dependencies with compatible versions
langgraph>=0.2.0,<0.7.0
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langsmith>=0.1.0,<0.2.0

# AWS and utilities
boto3>=1.34.0
pydantic>=2.8.0,<3.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
typing-extensions>=4.12.0
